+++ src/index.ts
@@ -3,3 +3,11 @@
 export * from './interfaces'
 export * from './util/ordered'
 export * from './util/map'
+
+export * from './exception/exception'
+export * from './exception/ex-file-system'
+export * from './exception/ex-illegal-access'
+export * from './exception/ex-not-implemented'
+export * from './exception/ex-out-of-bounds'
+export * from './exception/ex-sealed-access'
+export * from './exception/ex-unsupported-operation'