+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## DEVELOPMENT
 
+### 0.2.0
+`+` Exceptions, an extension of the base JS Error  
+`*` Minor changes to `ComparableMap` to support the new exceptions  
+
 #### 0.1.1
 `*` Correct typing exports